+++ MdeModulePkg/Universal/Network/Tcp4Dxe/Tcp4Misc.c
@@ -657,7 +657,7 @@ TcpResetConnection (
   Nhead->Ack      = HTONL (Tcb->RcvNxt);
   Nhead->SrcPort  = Tcb->LocalEnd.Port;
   Nhead->DstPort  = Tcb->RemoteEnd.Port;
-  Nhead->HeadLen  = (sizeof (TCP_HEAD) >> 2);
+  Nhead->HeadLen  = (UINT8) (sizeof (TCP_HEAD) >> 2);
   Nhead->Res      = 0;
   Nhead->Wnd      = HTONS (0xFFFF);
   Nhead->Checksum = 0;
